+++ core/http/httpext.py
@@ -2,8 +2,8 @@ import asyncio
 import typing
 import re
 import aiofiles
-from core.util import util, io, tasks
-from core.msg import msgabc, msgext, msgftr
+from core.util import util, io, tasks, aggtrf, funcutil
+from core.msg import msgabc, msgext, msgftr, msgtrf
 from core.http import httpabc, httpcnt, httpsubs
 
 
@@ -175,3 +175,30 @@ class _FileByteStream(httpabc.ByteStream):
                 pumping = chunk is not None and chunk != b''
         tasks.task_end(self._task)
         self._task = None
+
+
+class RollingLogHandler(httpabc.AsyncGetHandler):
+
+    def __init__(self, mailer: msgabc.MulticastMailer, msg_filter: msgabc.Filter, size: int = 100):
+        self._mailer = mailer
+        self._subscriber = msgext.RollingLogSubscriber(
+            mailer, size=size,
+            msg_filter=msg_filter,
+            transformer=msgtrf.GetData(),
+            aggregator=aggtrf.StrJoin('\n'))
+        mailer.register(self._subscriber)
+
+    async def handle_get(self, resource, data):
+        return await msgext.RollingLogSubscriber.get_log(self._mailer, self, self._subscriber.get_identity())
+
+
+class WipeHandler(httpabc.AsyncPostHandler):
+
+    def __init__(self, path: str, callback: typing.Callable = None):
+        self._path = path
+        self._callback = callback
+
+    async def handle_post(self, resource, data):
+        await io.delete_directory(self._path)
+        await funcutil.silently_call(self._callback)
+        return httpabc.ResponseBody.NO_CONTENT
