+++ servers/factorio/server.py
@@ -1,7 +1,6 @@
 from core.util import aggtrf
-from core.msg import msgabc, msgtrf, msgext
 from core.context import contextsvc
-from core.http import httpabc, httprsc, httpsubs
+from core.http import httpabc, httprsc, httpsubs, httpext
 from core.proc import proch, prcext
 from core.system import svrabc, svrsvc, svrext
 from servers.factorio import deployment as dep, messaging as msg, playerstore as pls
@@ -33,8 +32,8 @@ class Server(svrabc.Server):
             .append('subscribe', self._httpsubs.handler(msg.PLAYER_EVENT_FILTER)) \
             .pop() \
             .push('log') \
-            .append('tail', _ConsoleLogHandler(self._context)) \
-            .append('subscribe', self._httpsubs.handler(msg.CONSOLE_FILTER, aggtrf.StrJoin('\n'))) \
+            .append('tail', httpext.RollingLogHandler(self._context, msg.CONSOLE_LOG_FILTER)) \
+            .append('subscribe', self._httpsubs.handler(msg.CONSOLE_LOG_FILTER, aggtrf.StrJoin('\n'))) \
             .pop() \
             .push(self._httpsubs.resource(resource, 'subscriptions')) \
             .append('{identity}', self._httpsubs.subscriptions_handler('identity'))
@@ -49,18 +48,3 @@ class Server(svrabc.Server):
 
     async def stop(self):
         await self._stopper.stop()
-
-
-class _ConsoleLogHandler(httpabc.AsyncGetHandler):
-
-    def __init__(self, mailer: msgabc.MulticastMailer):
-        self._mailer = mailer
-        self._subscriber = msgext.RollingLogSubscriber(
-            mailer, size=100,
-            msg_filter=msg.CONSOLE_FILTER,
-            transformer=msgtrf.GetData(),
-            aggregator=aggtrf.StrJoin('\n'))
-        mailer.register(self._subscriber)
-
-    async def handle_get(self, resource, data):
-        return await msgext.RollingLogSubscriber.get_log(self._mailer, self, self._subscriber.get_identity())
